+++ packages/svelte-reveal/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New docs website
+
 ### Changed
 
 - Improved GitHub actions workflows ([#208](https://github.com/DaveKeehl/svelte-reveal/issues/208))