+++ packages/svelte-reveal/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Custom preprocessor to reduce boilerplate when using SvelteKit ([#222](https://github.com/DaveKeehl/svelte-reveal/pull/222))
+
 ## [1.1.0] - 2024-04-10
 
 ### Added