+++ packages/svelte-reveal/CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New docs website ([#211](https://github.com/DaveKeehl/svelte-reveal/issues/211))
 - Custom preprocessor to reduce boilerplate when using SvelteKit ([#222](https://github.com/DaveKeehl/svelte-reveal/pull/222))
 
+
+### Changed
+
+- Updated dependencies ([#223](https://github.com/DaveKeehl/svelte-reveal/pull/223))
+
 ## [1.1.0] - 2024-04-10
 
 ### Changed