+++ packages/svelte-reveal/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0] - 2026-05-02
 
 ### Added
 
@@ -229,3 +229,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.0.3]: https://github.com/DaveKeehl/svelte-reveal/compare/1.0.2...1.0.3
 [1.0.4]: https://github.com/DaveKeehl/svelte-reveal/compare/1.0.3...1.0.4
 [1.1.0]: https://github.com/DaveKeehl/svelte-reveal/compare/1.0.4...1.1.0
+[1.2.0]: https://github.com/DaveKeehl/svelte-reveal/compare/1.1.0...1.2.0